HpクラスとStatblockクラスにヘッダーファイルのインクルードガードを、追加することにより,ヘッダーファイルが複数回インクルードされるのを防ぐことができ,解決できました。

+++ stattypes.h
@@ -1,5 +1,2 @@
 #include <cstdint>
-
-// 0‚©‚ç255‚Ü‚Å‚Ì”ÍˆÍ‚Ì’l‚µ‚©•\Œ»‚Å‚«‚È‚¢
-// uint8_t‚Í•„†‚È‚µ®”
-typedef std::uint8_t stattype;
+typedef std::uint16_t stattype;
